fix: static-binding hooks pass local verification (28/28 doctest on =yes)
- builtin hook: drop redundant uint32_t cast (param is Variant::Type now)
- class-method loop: method_name declared before the static-first hook;
  hook guard is value-semantic (#if); stray #endif removed
- SConstruct: /bigobj for windows when static_binding=yes (the dispatch TU
  instantiates ~15k thunks and overflows the default COFF section count)
- verified locally: scons static_binding=yes full build + --jsb-run-tests
  => 28 test cases / 449 assertions all passing, zero errors

+++ src/runtime/bridge/jsb_primitive_bindings_reflect.cpp
@@ -747,7 +747,7 @@ public:
 				}
 
 #if JSB_WITH_STATIC_BINDINGS
-				if (const jsb::static_binding::ThunkFn sb_thunk = jsb::static_binding::find_builtin_thunk((uint32_t)TYPE, method_info.method.name, method_info.hash)) {
+				if (const jsb::static_binding::ThunkFn sb_thunk = jsb::static_binding::find_builtin_thunk(TYPE, method_info.method.name, method_info.hash)) {
 					if (method_info.is_static()) {
 						class_builder.Static().Method(member_name, sb_thunk);
 					} else {
